Cambios en dtos

+++ src/main/java/uy/edu/ort/malapata/modelo/clasesDeCalculo/CalculadorDescuentoTriple.java
@@ -2,7 +2,7 @@ package uy.edu.ort.malapata.modelo.clasesDeCalculo;
 import uy.edu.ort.malapata.modelo.interfaces.CalculadorDescuento;
 
 public class CalculadorDescuentoTriple implements CalculadorDescuento {
-    public double calcular(double montoApostado, int totalApostadoAlCaballo) {
+    public double calcular(double montoApostado, double totalApostadoAlCaballo) {
         return montoApostado * 1.5;
     }
 }
