+++ src/main/java/uy/edu/ort/malapata/modelo/clasesDeCalculo/CalculadorPremioSimple.java
@@ -2,7 +2,7 @@ package uy.edu.ort.malapata.modelo.clasesDeCalculo;
 import uy.edu.ort.malapata.modelo.interfaces.CalculadorPremio;
 
 public class CalculadorPremioSimple implements CalculadorPremio {
-    public double calcular(double montoApostado, int dividendo) {
+    public double calcular(double montoApostado, double dividendo) {
         return montoApostado * dividendo;
     }
 }
