+++ src/main/java/uy/edu/ort/malapata/modelo/clasesDeCalculo/CalculadorPremioSuper.java
@@ -2,7 +2,7 @@ package uy.edu.ort.malapata.modelo.clasesDeCalculo;
 import uy.edu.ort.malapata.modelo.interfaces.CalculadorPremio;
 
 public class CalculadorPremioSuper implements CalculadorPremio {
-    public double calcular(double montoApostado, int dividendo) {
+    public double calcular(double montoApostado, double dividendo) {
         double factor = (dividendo >= 2) ? 3 : 4;
         return montoApostado * dividendo * factor;
     }
